+++ cs/Markdown/TokenTypes.cs
@@ -1,0 +1,18 @@
+﻿namespace Markdown
+{
+    internal enum TokenType
+    {
+        Text,
+        Space,
+        Newline,
+        Sharp,
+        Object
+    }
+
+    internal enum TokenObjectType
+    {
+        Italic,
+        Strong,
+        Header
+    }
+}
